requirements.txt
discord.py
wikipedia
googletrans